add sync blocks cronjob
